"refactor(#22): 게임 페이지 이동"
