+++ src/pages/game/ResultPage.js
