+++ src/pages/game/VotePage.js
